backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
groq==0.5.0
sentence-transformers==2.6.1
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.6.4
